# backend/Services/PunchDataService/internal/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Collects raw punch logs from biometric devices, file uploads, manual entry, and read-only import from a public ZKTeco SQL Server. Stores normalised rows in PunchRecords with deduplication. Does not calculate attendance (see AttendanceService).",
        "title": "PunchDataService API",
        "contact": {
            "name": "EnterpriseERP Platform Team"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/punch-data/remote/collect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Read-only SELECT from remote CHECKINOUT; stores new rows in PunchRecords with deduplication (company + employee + device + punch time). Does not modify the remote database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "remote-collect"
                ],
                "summary": "Collect punches from public ZKTeco SQL Server",
                "parameters": [
                    {
                        "description": "Collect window",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.remoteCollectBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-github_com_enterprise-erp_punchdata_internal_collector_RemoteCollectResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/punch-data/remote/collect/histories": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated history of read-only imports from the public ZKTeco database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "remote-collect"
                ],
                "summary": "List remote collect runs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Company filter",
                        "name": "companyId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size (max 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-github_com_enterprise-erp_punchdata_internal_response_PagedResult-github_com_enterprise-erp_punchdata_internal_models_RemoteCollectHistory"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/punch-data/remote/collect/preview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Counts rows on public CHECKINOUT in the time window without importing. Defaults to the last 24 hours when from/to are omitted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "remote-collect"
                ],
                "summary": "Preview remote punch row count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RFC3339 start (default: now - 24h)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "RFC3339 end (default: now)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-internal_handlers_remoteCollectPreviewResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_enterprise-erp_punchdata_internal_collector.RemoteCollectResult": {
            "type": "object",
            "properties": {
                "duplicates": {
                    "type": "integer"
                },
                "history": {
                    "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_models.RemoteCollectHistory"
                },
                "inserted": {
                    "type": "integer"
                },
                "logFile": {
                    "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_models.PunchLogFile"
                },
                "remoteRows": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_models.PunchLogFile": {
            "type": "object",
            "properties": {
                "companyId": {
                    "type": "integer"
                },
                "contentType": {
                    "type": "string"
                },
                "deviceId": {
                    "type": "string"
                },
                "errorMessage": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "importBatchId": {
                    "type": "string"
                },
                "processedAt": {
                    "type": "string"
                },
                "rowCount": {
                    "type": "integer"
                },
                "sizeBytes": {
                    "type": "integer"
                },
                "sourceType": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "uploadedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_models.RemoteCollectHistory": {
            "type": "object",
            "properties": {
                "companyId": {
                    "type": "integer"
                },
                "completedAt": {
                    "type": "string"
                },
                "duplicates": {
                    "type": "integer"
                },
                "errorMessage": {
                    "type": "string"
                },
                "fromTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inserted": {
                    "type": "integer"
                },
                "logFileId": {
                    "type": "string"
                },
                "remoteRows": {
                    "type": "integer"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "toTime": {
                    "type": "string"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_response.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_enterprise-erp_punchdata_internal_response.ApiResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiError"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_response.ApiResponse-github_com_enterprise-erp_punchdata_internal_collector_RemoteCollectResult": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_collector.RemoteCollectResult"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiError"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_response.ApiResponse-github_com_enterprise-erp_punchdata_internal_response_PagedResult-github_com_enterprise-erp_punchdata_internal_models_RemoteCollectHistory": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.PagedResult-github_com_enterprise-erp_punchdata_internal_models_RemoteCollectHistory"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiError"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_response.ApiResponse-internal_handlers_remoteCollectPreviewResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_handlers.remoteCollectPreviewResponse"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_response.ApiError"
                    }
                },
                "success": {
                    "type": "boolean"
                },
                "traceId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_enterprise-erp_punchdata_internal_response.PagedResult-github_com_enterprise-erp_punchdata_internal_models_RemoteCollectHistory": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_enterprise-erp_punchdata_internal_models.RemoteCollectHistory"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handlers.remoteCollectBody": {
            "type": "object",
            "properties": {
                "batchSize": {
                    "type": "integer"
                },
                "companyId": {
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "useWatermark": {
                    "type": "boolean"
                }
            }
        },
        "internal_handlers.remoteCollectPreviewResponse": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "readOnly": {
                    "type": "boolean"
                },
                "remoteRows": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT issued by AuthService. Send as: `Bearer \u003ctoken\u003e`. Pass companyId in body or query when needed.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}